# openapi.yaml
openapi: 3.0.3
info:
  title: EMLy API
  description: |
    REST API for the EMLy bug reporting system.

    ## Authentication

    | Scheme | Header | Used for |
    |--------|--------|----------|
    | `ApiKey` | `X-API-Key` | Submitting and reading bug reports |
    | `AdminKey` | `X-Admin-Key` | User management, full bug-report access, update management |
    | `SessionToken` | `X-Session-Token` | Returned by `POST /v2/api/admin/auth/login`; used for session-validated requests |

    Endpoints that require both `ApiKey` and `AdminKey` list both schemes under `security`.

    ## Rate limiting

    All route groups: **30 req / min per IP**.
    Global ceiling: **100 req / min per IP**.

    ## Versioning

    - **v2** — current, recommended for all new integrations
    - **v1** — legacy; identical to v2 except bug-report path (`/bug-reports` plural) and no `/updates` group
  version: "2.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    SessionToken:
      type: apiKey
      in: header
      name: X-Session-Token

  # ─────────────────────────────────────────────────────────────
  # Shared schemas
  # ─────────────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "bug report not found"

    # ── Bug Reports ──────────────────────────────────────────
...
    ReleaseChannel:
      type: string
      enum: [stable, beta, archived]
      example: stable
    ReleaseSeverity:
      type: string
      enum: [none, security, bugfix, feature]
      example: none
    ReleaseProduct:
      type: string
      enum: [app, updater]
      description: >
        Which binary the release is for — the EMLy desktop app or the EMLy Updater.
        `channel` and `is_critical` are singleton slots scoped per product (e.g. there
        can be one `app` stable release and one `updater` stable release at the same time).
      example: app
...
    Release:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ReleaseProduct'
        version:
          type: string
          example: "1.7.0"
        channel:
          $ref: '#/components/schemas/ReleaseChannel'
        download_filename:
          type: string
          example: "EMLy_Installer_1.7.0.exe"
        sha256_checksum:
          type: string
          example: "abc123def456..."
        short_note:
          type: string
          example: "Cambiamento sistema aggiornamento."
        severity_type:
          $ref: '#/components/schemas/ReleaseSeverity'
        description_en:
          type: string
          nullable: true
          example: "Fixed a critical PDF rendering vulnerability."
        description_it:
          type: string
          nullable: true
          example: "Corretto un bug critico nel rendering PDF."
        is_critical:
          type: boolean
          example: false
        critical_version:
          type: string
          nullable: true
          description: Version clients must update to when `is_critical` is true; only one release can hold this at a time
          example: "1.7.0"
        min_required_version:
          type: string
          nullable: true
          example: "1.6.0"
        released_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

# ─────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────
paths:
...
  /v2/updates/releases:
    get:
      tags: [Updates]
      summary: List all releases
      description: Returns releases ordered by `released_at` descending. Filterable by channel.
      security:
        - AdminKey: []
      parameters:
        - name: channel
          in: query
          description: Filter by channel
          schema:
            $ref: '#/components/schemas/ReleaseChannel'
      responses:
        "200":
          description: Array of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Release'
        "401":
          description: Missing or invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Updates]
      summary: Create a new release and upload the installer
      description: |
        Accepts `multipart/form-data`. Uploads the `.exe` to R2/S3, computes
        the SHA-256 checksum server-side, and saves the release record.

        `channel` defaults to `archived`, `severity_type` defaults to `none`.
        If `severity_type` is not `none`, the release will appear in
        `detailedReleaseNotes` in the manifest.

        Returns `503` if S3 storage is not configured on the server.
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [version, file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The installer binary (.exe). Filename is stored as-is.
                version:
                  type: string
                  example: "1.7.5"
                channel:
                  $ref: '#/components/schemas/ReleaseChannel'
                short_note:
                  type: string
                  example: "Fix critico al renderer PDF."
                severity_type:
                  $ref: '#/components/schemas/ReleaseSeverity'
                description_en:
                  type: string
                  description: Omit or leave empty to exclude from detailed notes
                  example: "Fixed a critical PDF rendering vulnerability."
                description_it:
                  type: string
                  description: Omit or leave empty to exclude from detailed notes
                  example: "Corretto un bug critico nel rendering PDF."
                is_critical:
                  type: string
                  enum: ["true", "false", "1", "0"]
                  default: "false"
                  description: When `true`, clears the critical flag from any other release and sets this as the sole critical release
                critical_version:
                  type: string
                  description: The version clients must update to when `is_critical` is true
                  example: "1.7.5"
                min_required_version:
                  type: string
                  description: Clients below this version are forced to update; omit to leave unset
                  example: "1.6.0"
                released_at:
                  type: string
                  format: date-time
                  description: ISO 8601 timestamp; defaults to current UTC time if omitted
      responses:
        "201":
          description: Release created
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                    example: "1.7.5"
                  channel:
                    $ref: '#/components/schemas/ReleaseChannel'
                  download_filename:
                    type: string
                    example: "EMLy_Installer_1.7.5_beta.exe"
                  sha256_checksum:
                    type: string
                    example: "a3f1bc..."
        "400":
          description: Validation error or missing file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Missing or invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "503":
          description: S3 storage not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/updates/{product}/releases:
    get:
      tags: [Updates]
      summary: List all releases for a product
      description: Same as `GET /v2/updates/releases`, scoped to `{product}`. Filterable by channel.
      security:
        - AdminKey: []
      parameters:
        - name: product
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ReleaseProduct'
        - name: channel
          in: query
          description: Filter by channel
          schema:
            $ref: '#/components/schemas/ReleaseChannel'
      responses:
        "200":
          description: Array of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Release'
        "401":
          description: Missing or invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Unknown product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Updates]
      summary: Create a new release for a product
      description: |
        Same as `POST /v2/updates/releases`, scoped to `{product}`. Accepts
        `multipart/form-data` with the same fields (`version`, `file`, `channel`,
        `short_note`, `severity_type`, `description_en`, `description_it`,
        `is_critical`, `critical_version`, `min_required_version`, `released_at`).
        `channel`/`is_critical` singleton slots are scoped to `{product}`.
      security:
        - AdminKey: []
      parameters:
        - name: product
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ReleaseProduct'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [version, file]
              properties:
                file:
                  type: string
                  format: binary
                version:
                  type: string
                  example: "1.0.3"
                channel:
                  $ref: '#/components/schemas/ReleaseChannel'
                short_note:
                  type: string
                severity_type:
                  $ref: '#/components/schemas/ReleaseSeverity'
                description_en:
                  type: string
                description_it:
                  type: string
                is_critical:
                  type: string
                  enum: ["true", "false", "1", "0"]
                  default: "false"
                critical_version:
                  type: string
                min_required_version:
                  type: string
                released_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Release created
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/ReleaseProduct'
                  version:
                    type: string
                  channel:
                    $ref: '#/components/schemas/ReleaseChannel'
                  download_filename:
                    type: string
                  sha256_checksum:
                    type: string
        "400":
          description: Validation error or missing file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Missing or invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "503":
          description: S3 storage not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...